requirements.txt
asgiref==3.7.2
Django==3.2.10
djangorestframework==3.13.1
gunicorn==21.2.0
packaging==23.1
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.5.0
